Login - codeOrganization AuthModule organized
AuthModule was organized as follows:
Helpers folder:
- AuthGuard with CanActivate and CanLoad methods
- Auth interceptor for adding authentication header to http requests

Login component
- login form
- login high level logic

Models folder
- Required authentication models (TokenRequest Interface)

Register component
- register form
- register high level logic

Services folder
- Authentication service (contains social authService). Responsible of
all the authentication tasks.

AuthRoutingModule and AuthModule files
