+++ src/app/app-routing.module.ts
@@ -1,10 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './auth/login/login.component';
-import { DashboardComponent } from './user/dashboard/dashboard.component';
 import { AuthGuard } from './auth/helpers/auth.guard';
-import { AuthComponent } from './auth/auth/auth.component';
 
 
 
@@ -14,7 +11,7 @@ const routes: Routes = [
   },
   {
     path: 'user', loadChildren: () => import('./user/user.module').then(mod => mod.UserModule),
-    canActivate: [AuthGuard]
+    canLoad: [AuthGuard]
   },
   { path: 'home', component: HomeComponent },
   {
