+++ src/app/auth/auth-routing.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './login/login.component';
-import { AuthComponent } from './auth/auth.component';
 
 
 const authRoutes: Routes = [
