--- src/app/auth/auth/auth.component.html
@@ -1,2 +1,0 @@
-<router-outlet></router-outlet>
-
