+++ src/app/auth/login/login.component.ts
@@ -9,7 +9,7 @@ import { AuthenticationService } from '../services/authentication.service';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Observable, of } from 'rxjs';
 import { first } from 'rxjs/operators';
-import { LoggingService } from '@app/logging.service';
+import { LoggingService } from '@app/services/logging.service';
 import { SocialUser } from 'angularx-social-login';
 import { User } from '@app/user/models/user.model';
