+++ src/app/user/dashboard/dashboard.component.ts
@@ -3,7 +3,7 @@ import { User } from '../models/user.model';
 import { UserService } from '../services/user.service';
 import { Observable } from 'rxjs';
 import { logging } from 'protractor';
-import { LoggingService } from '@app/logging.service';
+import { LoggingService } from '@app/services/logging.service';
 import { AuthenticationService } from '@app/auth/services/authentication.service';
 import { Router } from '@angular/router';
